Refactor: @Data 수정
