+++ src/main/java/com/travelcompass/api/oauth/jwt/JwtRequestDto.java
@@ -1,8 +1,8 @@
 package com.travelcompass.api.oauth.jwt;
 
-import lombok.Data;
+import lombok.Getter;
 
-@Data
+@Getter
 public class JwtRequestDto {
     private String username;
     private String password;
